pyproject.toml
[project]
name = "xenty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kaggle>=1.7.4.5",
    "kagglehub>=0.3.12",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "pip>=25.1.1",
    "plotly>=6.2.0",
    "polars>=1.31.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "streamlit>=1.46.1",
    "tweepy>=4.16.0",
]
